endnote behavior is now cff-version specific; updated tests accordingly
