+++ cffconvert/behavior_1_2_x/citation.py
@@ -7,6 +7,7 @@ from cffconvert.root import get_package_root
 from cffconvert.behavior_1_2_x.apalike import ApalikeObject
 from cffconvert.behavior_1_2_x.bibtex import BibtexObject
 from cffconvert.behavior_1_2_x.codemeta import CodemetaObject
+from cffconvert.behavior_1_2_x.endnote import EndnoteObject
 from cffconvert.behavior_1_2_x.schemaorg import SchemaorgObject
 from cffconvert.behavior_1_2_x.zenodo import ZenodoObject
 
@@ -51,6 +52,9 @@ class Citation_1_2_x(Contract):
 
     def as_codemeta(self):
         return CodemetaObject(self.cffobj).print()
+
+    def as_endnote(self):
+        return EndnoteObject(self.cffobj).print()
 
     def as_schemaorg(self):
         return SchemaorgObject(self.cffobj).print()
