+++ cffconvert/contracts/citation.py
@@ -25,6 +25,10 @@ class Contract(ABC):
         pass
 
     @abstractmethod
+    def as_endnote(self):
+        pass
+
+    @abstractmethod
     def as_schemaorg(self):
         pass
